+++ javascript/reducers/currentFractal.js
@@ -1,6 +1,0 @@
-import createReducer from 'reducers/createReducer'
-
-export default createReducer('fractal', 'julia set')
-
-export const getCurrentFractal = (state) => state.currentFractal
-