+++ javascript/reducers/currentShader.js
@@ -1,0 +1,6 @@
+import createReducer from 'reducers/createReducer'
+
+export default createReducer('shader', 'julia set')
+
+export const getCurrentShader = (state) => state.currentShader
+