+++ shaders/burningShip.glsl
@@ -101,7 +101,7 @@ void main() {
 
   if (COLORSET == 0.0) {
     float color = BRIGHTNESS * fractalValue.x / float(MAX_ITERATIONS);
-    gl_FragColor = vec4(color, color, color, 0.0);
+    gl_FragColor = vec4(color, color, color, 1.0);
   } else if (COLORSET == 1.0) {
     gl_FragColor = BRIGHTNESS * colorize(fractalValue);
   }